Add sudo permissions when needed to allow getting non-root username [see more]
This commit eliminates the need to run the script 'run-scripts.sh' with sudo permissions, beacuse the permissions is asked when necessary. Otherwise (that is, executing 'sudo run-scripts' to enable installations), it wouldn't be possible to get the non-root username of the machine when needed (e.g. when using 'sed' to set 'zsh' as default shell).